requirements.txt
markdown-it-py==4.0.0
mdurl==0.1.2
paho-mqtt==2.1.0
Pygments==2.19.2
rich==14.2.0
